+++ cli/deps.ts
@@ -1,17 +1,17 @@
 export { z } from 'https://esm.sh/zod@3.20.2'
-export { Command } from 'https://deno.land/x/cliffy@v0.25.7/command/mod.ts'
+export { Command } from 'https://deno.land/x/cliffy@v1.0.0-rc.1/command/mod.ts'
 export {
   Input,
   prompt,
   Secret,
   Select,
   Toggle,
-} from 'https://deno.land/x/cliffy@v0.25.7/prompt/mod.ts'
+} from 'https://deno.land/x/cliffy@v1.0.0-rc.1/prompt/mod.ts'
 export {
   createClient,
   SupabaseClient,
 } from 'https://esm.sh/@supabase/supabase-js@2.4.1'
-export { wait } from 'https://deno.land/x/wait@0.1.12/mod.ts'
+export { Spinner, wait } from 'https://deno.land/x/wait@0.1.13/mod.ts'
 export { default as $ } from 'https://deno.land/x/dax@0.31.0/mod.ts'
 export { delay } from 'https://deno.land/std@0.179.0/async/mod.ts'
 export { join } from 'https://deno.land/std@0.179.0/path/mod.ts'